Made some changes to preprocessing and featurization scripts

+++ src/preprocessing.py
@@ -1,35 +1,32 @@
 from rdkit import Chem
-from rdkit.Chem import SaltRemove, AllChem
+from rdkit.Chem import AllChem
+import os
+import sys
 
-def standardize_smiles(smiles: str) -> str:
-  # Removing salts and neutralizing molecules in our dataset
-  try:
-    mol = Chem.MolFromSmiles(smiles)
-    if not mol:
-      return None
-
-    # remove salts\
-    remover = SaltRemover.SaltRemover()
-    mol = remover.StripMol(mol)
-
-    # change all molecules to canonical SMILES
-    return Chem.MolToSmiles(mol, canonical=True)
-
-  except:
-    return None
+project_root = os.path.dirname(os.path.abspath(__file__))
+os.chdir(project_root)
 
 def preprocess_data(input_csv: str, output_csv: str):
-  #Load the raw data, standardize SMILES, save clean data
+  #Load the bioactivity data, do some minor processing, and add an activity column for active or inactive
   import pandas as pd
+  print(f"Reading from: {input_csv}")
 
+  if not os.path.exists(input_csv):
+    raise FileNotFoundError(f"Input file not found: {input_csv}")  
+  
   df = pd.read_csv(input_csv)
-  df["standardized_smiles"] = df["smiles"].apply(standardize_smiles)
 
-  #Drop invalid rows
-  df = df.dropna(subset=["standardized_smiles"])
+  #Drop rows with missing values
+  df = df.dropna()
+
+  df['standard_value'] = pd.to_numeric(df['standard_value'], errors='coerce')
+  threshold=1000
+  df['bioactivity_class'] = df['standard_value'].apply(lambda x: 'active' if x < threshold else 'inactive')
 
   df.to_csv(output_csv, index=False)
   print(f"Data has been preprocessed and saved to {output_csv}")
 
 if __name__ == "__main__":
-  preprocess_data("data/raw/egfr_bioactivity_full.csv", "data/processed/egfr_cleaned.csv")
+  input_csv = os.path.abspath("../data/raw/egfr_bioactivity_full.csv")
+  output_csv = os.path.abspath("../data/processed/egfr_cleaned.csv")
+  preprocess_data(input_csv, output_csv)
